+++ OWLOS/src/drivers/LightDriver.cpp
@@ -91,7 +91,7 @@ bool LightDriver::publish()
 	return false;
 };
 
-String LightDriver::onMessage(String _topic, String _payload, int transportMask)
+String LightDriver::onMessage(String _topic, String _payload, int8_t transportMask)
 {
 	String result = BaseDriver::onMessage(_topic, _payload, transportMask);
 	if (!available) return result;
