+++ OWLOS/src/drivers/OptoDriver.cpp
@@ -100,7 +100,7 @@ bool OptoDriver::publish()
 	return false;
 };
 
-String OptoDriver::onMessage(String _topic, String _payload, int transportMask)
+String OptoDriver::onMessage(String _topic, String _payload, int8_t transportMask)
 {
 	String result = BaseDriver::onMessage(_topic, _payload, transportMask);
 	if (!available) return result;
