+++ OWLOS/src/drivers/SmokeDriver.cpp
@@ -89,7 +89,7 @@ bool SmokeDriver::publish()
 	return false;
 }
 
-String SmokeDriver::onMessage(String _topic, String _payload, int transportMask)
+String SmokeDriver::onMessage(String _topic, String _payload, int8_t transportMask)
 {
 	String result = BaseDriver::onMessage(_topic, _payload, transportMask);
 	if (!available) return result;
