+++ OWLOS/src/drivers/StepperDriver.cpp
@@ -98,7 +98,7 @@ String StepperDriver::getAllProperties()
 	return result;
 }
 
-String StepperDriver::onMessage(String _topic, String _payload, int transportMask)
+String StepperDriver::onMessage(String _topic, String _payload, int8_t transportMask)
 {
 	String result = BaseDriver::onMessage(_topic, _payload, transportMask);
 	if (!available) return result;
